refactor: move consolidated result processing to its own file
